text renderer, icc profile parser
both are not fully complete, but the text renderer works. it's from the
wx demo (which is itself from somewhere else, credited in comments).

I included some ascii tables because I was originally going to do that
(render a bitmap font), but the wxDC stuff seems probably better, at
least for now.